recup et ajout bdd marche

+++ app/src/main/java/com/example/td_to_do/dao/TodoDAO.java
@@ -45,7 +45,7 @@ public class TodoDAO extends DAO {
         if(cursor.moveToFirst()){
             while (!cursor.isAfterLast()){
                 Todo todo = new Todo();
-                todo.setId(cursor.getLong(TodoDBHelper.TODO_KEY_COLUMN_INDEX));
+                todo.setId((int) cursor.getLong(TodoDBHelper.TODO_KEY_COLUMN_INDEX));
                 todo.setName(cursor.getString(TodoDBHelper.TODO_NAME_COLUMN_INDEX));
                 todo.setUrgency(cursor.getString(TodoDBHelper.TODO_URGENCY_COLUMN_INDEX));
